+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 
 ## Upcoming Changes
 
+## TShock 4.3.21
+* Compatibility with Terraria 1.3.4.3 (@Patrikkk, @Zaicon).
+* API: Version tick 1.26.
+* API: Deprecated PlayerDamage and PlayerKillMe packets (now uses PlayerHurtV2 and PlayerDeathV2).
+* API: Main.rand now uses UnifiedRandom instead of Random. This WILL break any existing plugin that uses Main.rand.
 * Fixed HealOtherPlayer packet exploit (@Simon311).
 * Added associated config option for HealOtherPlayer exploit prevention (@Simon311).
 * Added `/accountinfo` command to get account information for a given TShock account (@Simon311).